code/pipeline/wsd_SR/requirements.txt
pytorch_lightning==0.8.5
transformers==2.10.0
nltk
six
torch-scatter==2.0.5 -f https://pytorch-geometric.com/whl/torch-1.5.0+${CUDA}.html
torch-sparse==0.6.5 -f https://pytorch-geometric.com/whl/torch-1.5.0+${CUDA}.html 